+++ mycodo/config.py
@@ -56,6 +56,7 @@ DEVICES = [
     ('DS18B20', 'Temperature: DS18B20 (1-Wire)'),
     ('ATLAS_PT1000_I2C', 'Temperature: Atlas Scientific PT-1000 (I2C)'),
     ('ATLAS_PT1000_UART', 'Temperature: Atlas Scientific PT-1000 (Serial)'),
+    ('MAX31855', 'Temperature: MAX31855 with k-type thermocouple (Serial)'),
     ('TMP006', 'Temperature (Contactless): TMP 006/007 (I2C)'),
     ('AM2315', 'Temperature/Humidity: AM2315 (I2C)'),
     ('DHT11', 'Temperature/Humidity: DHT11 (GPIO)'),
@@ -149,6 +150,9 @@ DEVICE_INFO = {
     'K30_UART': {
         'py-dependencies': ['serial'],
         'measure': ['co2']},
+    'MAX31855': {
+        'py-dependencies': ['Adafruit_MAX31855'],
+        'measure': ['temperature', 'temperature_die']},
     'MH_Z16_I2C': {
         'i2c-addresses': ['0x63'],
         'i2c-address-change': False,
@@ -398,6 +402,7 @@ LIST_DEVICES_I2C = [
 
 # Devices that use SPI to communicate
 LIST_DEVICES_SPI = [
+    'MAX31855',
     'MCP3008'
 ]